Load data for NORIC

Functions for loading data needed by NORIC from a database

getAp(registryName, fromDate, toDate, singleRow, singleHospital = NULL, ...)

getSo(registryName, fromDate, toDate, singleRow, ...)

getAk(registryName, fromDate, toDate, singleRow, ...)

getFo(registryName, fromDate, toDate, singleRow, ...)

getAnP(registryName, fromDate, toDate, singleRow, singleHospital = NULL, ...)

getCt(registryName, fromDate, toDate, singleRow, ...)

getAkOppf(registryName, fromDate, toDate, singleRow, ...)

getAnD(registryName, fromDate, toDate, singleRow, singleHospital = NULL, ...)

getSs(registryName, fromDate, toDate, singleRow, singleHospital = NULL, ...)

getMk(registryName, fromDate, toDate, singleRow, ...)

getPs(registryName, fromDate, toDate, singleRow, ...)

getApLight(registryName, fromDate, toDate, singleRow, ...)

getTaviProm(registryName, fromDate, toDate, singleRow, ...)

Arguments

registryName

Character string defining the registry name.

fromDate

Character string of format YYYY-MM-DD with start date. Value NULL if no filter on date.

toDate

Character string of format YYYY-MM-DD with end date. Value NULL if no filter on date.

singleRow

Logical if only one row from the table is to be provided. Default value is FALSE.

singleHospital

if only data from one hospital, when national database. Default value is NULL, contains reshID of selected hospital else.

...

Optional arguments to be passed to the function.

Value

Data frame or (when multiple data sets are returned) a list of data frames containing registry data. In case of getNameReshId() data may also be returned as a named list of values (see Details).

Details

The functions getAp, getSs, getAk etc. load the AP-, SS- and AK- tables etc. respectively. For most tables, the query adds selected variables from ForlopsOversikt.