Liggedogn

Antall dager fra AnkomstPCIDato til Utskrivingsdato, kun for primaerforlop hos direkte innlagte pasienter eller pasienter overfoert fra annet sykehus.

legg_til_liggedogn(df_ap)

Arguments

df_ap
data.frame med Angio PCI data fra noric. Maa inneholde variablene AnkomstPCIDato, UtskrivingsDato, OverflyttetFra og Regtype.

Value

Returnerer df_ap med to nye variabler liggedogn og liggedogn_data. Variabelen liggedogn inneholder antall dager mellom AnkomstPCI-sykehus og utskrivelse, denne kan ogsaa inneholde negative tider og tider over 60 dager. Variabelne liggetid_dg inneholder datagrunnlaget for liggedogn og har verdiene:

"nei" for pasienter overfoert fra annen avdeling paa sykehuset eller der denne informasjonen er manglende.

"nei" for sekundaerforlop, da disse ikke har utskrivelsesskjema.

"ja" for primaerforlop hos direkte innlagte pasienter eller pasienter overfoert fra andre sykehus og der liggedogn er i intervallet 0 dager til 60 dager.

"ugyldig" for primaerforlop hos direkte innlagte pasienter eller pasienter overfoert fra andre sykehusmed og der liggetid er enten et negativt antall dager eller over 60 dager.

"manglende" for primaerforlop hos direkte innlagte pasienter eller pasienter overfoert fra andre sykehus der utskrivingsdato og/eller AnkomstPCIDato mangler.