Add time-variables

The function legg_til_tidsvariabler() creates six time-variables, aar, maaned_nr, maaned, kvartal, uke, aar_uke, based on a given date-column (var) of input data.frame df. Then these new variables are added to df before the entire data.frame is returned.

legg_til_tidsvariabler(df, var = ProsedyreDato)

Arguments

df
data.frame from any NORIC-table. Must contain a variable in date-format.

var
name of variable in date-format that should be used to calculate the new variables.

Value

The input data.frame df is returned as it is, with six new columns added,one for each time-variable.

Examples

df <- data.frame(
   datoVariabel = c("2021-02-28",
                    "2013-09-15",
                    "2020-05-31",
                    "2015-12-12",
                    "2016-12-31",
                    "2016-01-01" ))
legg_til_tidsvariabler(df = df,
                       var = datoVariabel)

#>   datoVariabel  aar maaned_nr  maaned kvartal uke aar_uke
#> 1   2021-02-28 2021        02 2021-02 2021-Q1  08 2021-08
#> 2   2013-09-15 2013        09 2013-09 2013-Q3  37 2013-37
#> 3   2020-05-31 2020        05 2020-05 2020-Q2  22 2020-22
#> 4   2015-12-12 2015        12 2015-12 2015-Q4  50 2015-50
#> 5   2016-12-31 2016        12 2016-12 2016-Q4  52 2016-52
#> 6   2016-01-01 2016        01 2016-01 2016-Q1  53 2015-53

if (FALSE) {
# FOR NORIC's ap, ak and ct-tables:
legg_til_tidsvariabler(ap, var = ProsedyreDato)
legg_til_tidsvariabler(ak, var = ProsedyreDato)
legg_til_tidsvariabler(ct, var = UndersokDato)
}