Hjelpefunksjoner for PCI-prom

For forløp der alle andre kriterier (alder, prosedyretype, indikasjon, utskrivelse, etc.) er oppfylt, det vil si kriterie_alle_midlertidig = ja. Sjekker om pasientene hadde eksisterende TAVI-PROM eller PCI-PROM det siste året. Sjekker om pasienten har fått TAVI i de 3 påfølgende månedene.

Legger til ny variabel:

kriterie_taviprom_siste_aar()

kriterie_ingen_taviprom har verdien har verdien nei dersom pasienten hadde TAVI-PROM det siste året.

kriterie_pciprom_siste_aar()

kriterie_ingen_pciprom har verdien har verdien nei dersom pasienten hadde PCI-PROM det siste året.

#' kriterie_ingen_tavi_neste3mnd()

kriterie_ingen_ny_tavi har verdien har verdien nei dersom pasienten har vært til en TAVI-prosedyre i etterkant av PCI prosedyren (0-3mnd).

kriterie_taviprom_siste_aar(df, registryName = NULL)

kriterie_pciprom_siste_aar(df, registryName = NULL)

kriterie_ingen_tavi_neste3mnd(df, registryName = NULL)

Arguments

df

Inneholder data fra pci + pciprom. Må inneholde variabelen kriterie_alle_midlertidig.

registryName

for SQL queries