NORIC staging data functions

NORIC's handling of staging data at Rapporteket.

makeStagingDataKi(registryName, rendered_by_shiny = FALSE)

makeStagingDataFrame(registryName)

checkValidStagingData(registryName, diffDaysCheck = 0)

deleteOldStagingData(registryName, diffDaysDelete)

bulletinProcessorStaging(
  dataset = "ki",
  orgName = "unknown organization",
  orgId = 999999,
  registryName = "noric",
  userFullName = "unknown person name",
  userRole = "unknown role",
  userOperator = "unknown operator",
  rendered_by_shiny = FALSE,
  author = "ingen"
)

Arguments

registryName

Character string defining the registry name.

rendered_by_shiny

boolean. if TRUE progression of pdf-generation is returned.

diffDaysCheck

numerical. Default values is 0 (today). checkValidStagingData() checks if any staging data has been created within the diffDaysCheck days.

diffDaysDelete

numerical. No default value. deleteOldStagingData() deletes staging data older than this.

dataset

Which kind of staging data to create

orgName

Character string with the name of the organization/hospital. Default is "unknown organization".

orgId

Integer (?) with the id of the organization/hospital. Default is 999999.

userFullName

Character string giving the person name, normally the user requesting the report. Default is "unknown person name".

userRole

Character string giving a user role, normally the one of the user requesting the report. Default is "unknown role".

userOperator

Character string with some name of an operator, whatever that is... Default is "unknown operator".

author

"ukjent"

Details

makeStagingDataKi() creates staging data for the KI-report.

makeStagingDataFrame() lists all of the existing staging data.

checkValidStagingData() checks whether any staging data set has been created within the last diffDaysCheck days. Returns valid_staging_data is TRUE and the name of the most recently created data set nyeste_staging_data if this is the case, and FALSE elsewhere.

deleteOldStagingData() deletes staging data older than diffDaysDelete

bulletinProcessorStaging() is a modified version of reportProcessor(). The function is applied to create bulletins for staged data.