Add variable aldersklasse in decades

Groups together ages in age-classes of decades, p.ex. "50-59" where ages 50 and 59 are min and max values of the age-class. Ages from 18 to 49 are grouped together, whereas ages under 18 years or above 99 years are given age-class <NA>.

utlede_aldersklasse(df, var = PasientAlder)

Arguments

df

data.frame, must contain a column with age as a continuous variable

var

is the variable name of the age-variable in df. Default value is PasientAlder.

Value

The input data.frame df is returned as it is, with one new column named aldersklasse added.

Examples

df <-  data.frame(age = c(10,20,59,60, NA, 61,69,70,99,100))
utlede_aldersklasse(df = df, var = age)
#>    age aldersklasse
#> 1   10         <NA>
#> 2   20        18-49
#> 3   59        50-59
#> 4   60        60-69
#> 5   NA         <NA>
#> 6   61        60-69
#> 7   69        60-69
#> 8   70        70-79
#> 9   99        90-99
#> 10 100         <NA>