New variables from NORIC's AnnenDiag-table

Functions for creating variables based on NORIC's AnnenDiag-table and for merging these variables into NORIC's AngioPCIVar-table by key variables ForlopsID and AvdRESH.

utlede_kar_annen_diag(df_ad = ad)

legg_til_trykk_bilde_per_kar(df_ap, df_ad, metodeType = "FFR")

Arguments

df_ad
NORIC's AnnenDiag-table, must contain variables ForlopsID, AVdRESH, segment, graft and metode

df_ap
NORIC's AngioPCIVar-table, must contain variables ForlopsID and AVdRESH

metodeType
value of variable metode from NORIC-table AnnenDiag. Must have one of following values : "FFR", "iFR", "IVUS" or "OCT"

Details

The function utlede_kar_annen_diag() groups the segments in df_ad into coronary arteries in the new variable kar, then df_ad is returned from the function.

The function legg_til_trykk_bilde_per_kar() uses the data in df_ad to calculate 5 new variables, where variable names prefix is type of method (one of: "FFR", "iFR", "IVUS" or "OCT") and suffix is each level of kar (p.ex. if metode is 'FFR' then variables are FFR_LAD, FFR_RCA, FFR_CX and FFR_Graft). The variable says whether or not a measure of this method is performed in each kar. If at least one measure is done in a kar the value of the corresponding variable is given the value ja, otherwise the value is nei. These 5 variables are then merged into df_ap, before df_ap is returned from the function. For procedures where no measures are done on segment-level (e.g. no entries available in df_ad), the values for the 5 new variables are <NA>.

Examples

df_ad <- data.frame(
   ForlopsID = c(1, 1, 3, 3, 5),
   AvdRESH = rep(1,5),
   segment = c("PDA/RPD (4)",
               "PLA (18)",
               "Midtre LAD (7)",
               "Ve hovedstamme (5)",
               "Ve hovedstamme (5)"),
   graft = c("Nei",
             "Arteriell",
             "Vene",
             "Nei",
             "Nei"),
   metode = c("FFR", "FFR", NA, "test", "FFR"))

df_ap <-data.frame(
   ForlopsID = c(1, 2, 3, 4, 5),
   AvdRESH = rep(1,5))

utlede_kar_annen_diag(df_ad)

#>   ForlopsID AvdRESH            segment     graft metode   kar
#> 1         1       1        PDA/RPD (4)       Nei    FFR   RCA
#> 2         1       1           PLA (18) Arteriell    FFR   RCA
#> 3         3       1     Midtre LAD (7)      Vene   <NA> Graft
#> 4         3       1 Ve hovedstamme (5)       Nei   test   LMS
#> 5         5       1 Ve hovedstamme (5)       Nei    FFR   LMS

legg_til_trykk_bilde_per_kar(df_ap = df_ap,
                             df_ad = df_ad,
                             metodeType = "FFR")

#>   ForlopsID AvdRESH FFR_LMS FFR_LAD FFR_RCA FFR_CX FFR_Graft
#> 1         1       1     nei     nei      ja    nei       nei
#> 2         2       1    <NA>    <NA>    <NA>   <NA>      <NA>
#> 3         3       1     nei     nei     nei    nei       nei
#> 4         4       1    <NA>    <NA>    <NA>   <NA>      <NA>
#> 5         5       1      ja     nei     nei    nei       nei