Add binary variable for table-status

The function utlede_ferdigstilt() uses the values of variable var to create a new column named ferdigstilt_suffix in df before df is returned from the function. The new variable is given the value ja if var is 1, and value nei if var is -1 or 0.

utlede_ferdigstilt(df, var = .data$SkjemaStatusStart, suffix = "startSkjema")

Arguments

df

data.frame, must contain a numerical status-variable with values -1, 0 or 1.

var

variable name of the numerical status-variable. Default value is SkjemaStatusStart.

suffix

wanted suffix in the new variable's name: ferdigstilt_suffix. Default values is "startSkjema".

Value

The data.frame given in input df is returned as it is, with one new column added.

Examples

x <- data.frame(SkjemaStatusStart = c(-1,-1,0,0,1,-1,0,1,NA),
                SkjemastatusHovedskjema = rep(0,9),
                SkjemaStatusUtskrivelse = rep(-1,9),
                SkjemaStatusKomplikasjoner = c(rep(-1,4), rep(1,3), NA, NA))

utlede_ferdigstilt(x, var = SkjemaStatusStart, suffix = "startSkjema")
#>   SkjemaStatusStart ferdigstilt_startSkjema SkjemastatusHovedskjema
#> 1                -1                     nei                       0
#> 2                -1                     nei                       0
#> 3                 0                     nei                       0
#> 4                 0                     nei                       0
#> 5                 1                      ja                       0
#> 6                -1                     nei                       0
#> 7                 0                     nei                       0
#> 8                 1                      ja                       0
#> 9                NA                    <NA>                       0
#>   SkjemaStatusUtskrivelse SkjemaStatusKomplikasjoner
#> 1                      -1                         -1
#> 2                      -1                         -1
#> 3                      -1                         -1
#> 4                      -1                         -1
#> 5                      -1                          1
#> 6                      -1                          1
#> 7                      -1                          1
#> 8                      -1                         NA
#> 9                      -1                         NA

utlede_ferdigstilt(x,
                   var = SkjemaStatusStart,
                   suffix = "startSkjema") %>%
   utlede_ferdigstilt(.,
                      var = SkjemastatusHovedskjema,
                      suffix = "hovedSkjema") %>%
   utlede_ferdigstilt(.,
                      var = SkjemaStatusKomplikasjoner,
                      suffix = "komplikSkjema")
#>   SkjemaStatusStart ferdigstilt_startSkjema SkjemastatusHovedskjema
#> 1                -1                     nei                       0
#> 2                -1                     nei                       0
#> 3                 0                     nei                       0
#> 4                 0                     nei                       0
#> 5                 1                      ja                       0
#> 6                -1                     nei                       0
#> 7                 0                     nei                       0
#> 8                 1                      ja                       0
#> 9                NA                    <NA>                       0
#>   ferdigstilt_hovedSkjema SkjemaStatusUtskrivelse SkjemaStatusKomplikasjoner
#> 1                     nei                      -1                         -1
#> 2                     nei                      -1                         -1
#> 3                     nei                      -1                         -1
#> 4                     nei                      -1                         -1
#> 5                     nei                      -1                          1
#> 6                     nei                      -1                          1
#> 7                     nei                      -1                          1
#> 8                     nei                      -1                         NA
#> 9                     nei                      -1                         NA
#>   ferdigstilt_komplikSkjema
#> 1                       nei
#> 2                       nei
#> 3                       nei
#> 4                       nei
#> 5                        ja
#> 6                        ja
#> 7                        ja
#> 8                      <NA>
#> 9                      <NA>