NORIC's quality indicators

Functions for creating two new variables for each indicator: The first variable says whether the procedure counts (ja/nei) in the denominator for the given indicator and the second says whether the procedure counts in the numerator (ja/nei/NA). If the denominator has value nei the numerator should always be NA. Denominator variables names are suffixed by _data (datagrunnlag).

ki_ferdigstilt_komplikasjoner(df_ap)

ki_trykkmaaling_utfoert(df_ap)

ki_ivus_oct_ved_stenting_lms(df_ap)

ki_foreskr_blodfortynnende(df_ap)

ki_foreskr_kolesterolsenkende(df_ap)

ki_nstemi_utredet_innen24t(df_ap)

ki_nstemi_utredet_innen72t(df_ap)

ki_stemi_pci_innen120min(df_ap)

Arguments

df_ap
NORIC's AngioPCIVar-table. Depending on indicators, must contain some of the variables SkjemaStatusKomplikasjoner, SkjemaStatusUtskrivelse, Regtype, Indikasjon, FFR, IFR, satt_inn_stent_i_LMS, TidlABC, IVUS, OCT, antall_stent_under_opphold, ...

Details

Following variables are created for the different functions:

ki_ferdigstilt_komplikasjoner()

denominator indik_komplik_ferdig_data (datagrunnlag) is ja when Regtype = Primaer.

numerator indik_komplik_ferdig has value ja if the complications form is completed (komplikasjons-skjema er ferdigstilt).

ki_trykkmaaling_utfoert()

denominator indik_trykkmaaling_data (datagrunnlag) is ja when Indikasjon = Stabil koronarsykdom.

numerator indik_trykkmaaling has value ja if FFR and/or iFR has been performed.

ki_ivus_oct_ved_stenting_lms()

denominator indik_ivus_oct_v_stent_lms_data (datagrunnlag) is ja when all of these conditions are fulfilled:

Indikasjon is one of "Vitieutredning", "Uklare brystsmerter", "Annet", "Hjertestans uten STEMI", "Hjertesvikt/kardiomyopati", "Komplettering av tidligere PCI", "UAP", "NSTEMI", "Stabil koronarsykdom"

satt_inn_stent_i_LMS = "ja"

TidlABC is one of "Nei", "Ukjent", NA

numerator indik_ivus_oct_v_stent_lms has value ja if IVUS and/or OCT has been performed.

ki_foreskr_blodfortynnende()

denominator indik_blodfortynnende_data (datagrunnlag) is ja when all of these conditions are fulfilled:

antall_stent_under_opphold is at least 1

Regtype = "Primær"

SkjemaStatusUtskrivelse is 1 (ferdigstilt)

UtskrevetDod = "Nei". Which means that "NA", "Ja" or "Ukjent" are excluded

numerator indik_blodfortynnende has value ja if at least two of ASA, AndrePlatehemmere or Antikoagulantia are prescribed.

ki_foreskr_kolesterolsenkende()

denominator indik_kolesterolsenkende_data (datagrunnlag) is ja when all of these conditions are fulfilled:

antall_stent_under_opphold is at least 1

Regtype = "Primær"

SkjemaStatusUtskrivelse is 1 (ferdigstilt)

UtskrevetDod = "Nei". Which means that "NA", "Ja" or "Ukjent" are excluded

numerator indik_kolesterolsenkende has value ja if UtskrStatiner is "Ja".

ki_nstemi_utredet_innen24t()

denominator indik_nstemi_angio_innen24t_data (datagrunnlag) is ja when all of these conditions are fulfilled:

Indikasjone is "NSTEMI"

Regtype = "Primær"

Innkomstarsak not "Øvrig" (NA allowed)

Hastegrad = "Akutt" or "Subakutt.

OverflyttetFra neither "Annen avdeling på sykehuset" nor "NA".

numerator indik_nstemi_angio_innen24t has value ja if ventetid_nstemi_timer is in the interval 0-24hours, value nei if ventetid_nstemi_timer is in the interval 24hours to 14 days and value ugyldig/manglende if time is negative, longer than 14 days or missing. See also function legg_til_ventetid_nstemi_timer.

ki_stemi_pci_innen120min()

denominator indik_stemi_pci_innen2t_data (datagrunnlag) is ja when all of these conditions are fulfilled:

AvdRESH is different from 106944 (AHUS Gardermoen does not treat STEMI patients)

Regtype = "Primær"

Indikasjon = "STEMI"

GittTrombolyse is "Nei" or "NA"

Hastegrad = "Akutt"

HLRForSykehus different from "Ja", "Ukjent"

ProsedyreType different from "Angio"

numerator indik_stemi_pci_innen2t has value ja if ventetid_stemi_min is in the interval 0-120 minutes, value nei if ventetid_stemi_min is in the interval 120 min to 24h (86400 min) and value ugyldig/manglende if time is negative, longer than 24h, missing or if BeslutningsutlosendeEKG has value "Prehospitalt" and ventetid_stemi_min is 0 minutes. See also function legg_til_ventetid_stemi_min.

Examples

 x <- data.frame(
      SkjemaStatusKomplikasjoner = c(-1, 1, 0, NA, NA, NA))
 noric::ki_ferdigstilt_komplikasjoner(df_ap = x)

#>   SkjemaStatusKomplikasjoner indik_komplik_ferdig_data indik_komplik_ferdig
#> 1                         -1                        ja                  nei
#> 2                          1                        ja                   ja
#> 3                          0                        ja                  nei
#> 4                         NA                       nei                 <NA>
#> 5                         NA                       nei                 <NA>
#> 6                         NA                       nei                 <NA>

 x <- data.frame(
      Indikasjon = c(rep("Stabil koronarsykdom", 4), NA, "annet"),
      FFR = c("Ja", "Ja", NA, "Ukjent", "Nei", "Ja"),
      IFR = c("Ja", "Nei", "Ukjent", NA, NA, NA))
 noric::ki_trykkmaaling_utfoert(df_ap = x)

#>             Indikasjon    FFR    IFR indik_trykkmaaling_data indik_trykkmaaling
#> 1 Stabil koronarsykdom     Ja     Ja                      ja                 ja
#> 2 Stabil koronarsykdom     Ja    Nei                      ja                 ja
#> 3 Stabil koronarsykdom   <NA> Ukjent                      ja                nei
#> 4 Stabil koronarsykdom Ukjent   <NA>                      ja                nei
#> 5                 <NA>    Nei   <NA>                     nei               <NA>
#> 6                annet     Ja   <NA>                     nei               <NA>

 x <- data.frame(
      Indikasjon = c(rep("Stabil koronarsykdom", 4), NA, "Annet"),
      TidlABC = rep("Nei", 6),
      satt_inn_stent_i_LMS = c(rep("ja", 4), NA, "nei"),
      IVUS = c("Ja", "Ja", NA, "Ukjent", "Nei", "Ja"),
      OCT = c("Ja", "Nei", "Ukjent", NA, NA, NA))
 noric::ki_ivus_oct_ved_stenting_lms(df_ap = x)

#>             Indikasjon TidlABC satt_inn_stent_i_LMS   IVUS    OCT
#> 1 Stabil koronarsykdom     Nei                   ja     Ja     Ja
#> 2 Stabil koronarsykdom     Nei                   ja     Ja    Nei
#> 3 Stabil koronarsykdom     Nei                   ja   <NA> Ukjent
#> 4 Stabil koronarsykdom     Nei                   ja Ukjent   <NA>
#> 5                 <NA>     Nei                 <NA>    Nei   <NA>
#> 6                Annet     Nei                  nei     Ja   <NA>
#>   indik_ivus_oct_v_stent_lms_data indik_ivus_oct_v_stent_lms
#> 1                              ja                         ja
#> 2                              ja                         ja
#> 3                              ja                        nei
#> 4                              ja                        nei
#> 5                             nei                       <NA>
#> 6                             nei                       <NA>

 x <- data.frame(
      antall_stent_under_opphold = 1:6,
      Regtype = c("Primær", "Primær", "Sekundær", rep("Primær", 3)),
      SkjemaStatusUtskrivelse = 1,
      UtskrevetDod = c(rep("Nei", 5), "Ja"),
      ASA = c("Ja", "Nei", "Ukjent", "Ja", "Ja", "Nei"),
      AndrePlatehemmere = c(NA, "Annet", "Clopidogrel (Plavix)",
                            "Nei", "Nei", NA),
      Antikoagulantia = c(NA, "Annet", "Lixiana", "Marevan",  NA, "Exanta"),
      UtskrStatiner = c("Ja", "Ja", NA, "Nei", "Ukjent", "Ja"))
noric::ki_foreskr_blodfortynnende(df_ap = x)

#>   antall_stent_under_opphold  Regtype SkjemaStatusUtskrivelse UtskrevetDod
#> 1                          1   Primær                       1          Nei
#> 2                          2   Primær                       1          Nei
#> 3                          3 Sekundær                       1          Nei
#> 4                          4   Primær                       1          Nei
#> 5                          5   Primær                       1          Nei
#> 6                          6   Primær                       1           Ja
#>      ASA    AndrePlatehemmere Antikoagulantia UtskrStatiner
#> 1     Ja                 <NA>            <NA>            Ja
#> 2    Nei                Annet           Annet            Ja
#> 3 Ukjent Clopidogrel (Plavix)         Lixiana          <NA>
#> 4     Ja                  Nei         Marevan           Nei
#> 5     Ja                  Nei            <NA>        Ukjent
#> 6    Nei                 <NA>          Exanta            Ja
#>   indik_blodfortynnende_data indik_blodfortynnende
#> 1                         ja                   nei
#> 2                         ja                    ja
#> 3                        nei                  <NA>
#> 4                         ja                    ja
#> 5                         ja                   nei
#> 6                        nei                  <NA>

noric::ki_foreskr_kolesterolsenkende(df_ap = x)

#>   antall_stent_under_opphold  Regtype SkjemaStatusUtskrivelse UtskrevetDod
#> 1                          1   Primær                       1          Nei
#> 2                          2   Primær                       1          Nei
#> 3                          3 Sekundær                       1          Nei
#> 4                          4   Primær                       1          Nei
#> 5                          5   Primær                       1          Nei
#> 6                          6   Primær                       1           Ja
#>      ASA    AndrePlatehemmere Antikoagulantia UtskrStatiner
#> 1     Ja                 <NA>            <NA>            Ja
#> 2    Nei                Annet           Annet            Ja
#> 3 Ukjent Clopidogrel (Plavix)         Lixiana          <NA>
#> 4     Ja                  Nei         Marevan           Nei
#> 5     Ja                  Nei            <NA>        Ukjent
#> 6    Nei                 <NA>          Exanta            Ja
#>   indik_kolesterolsenkende_data indik_kolesterolsenkende
#> 1                            ja                       ja
#> 2                            ja                       ja
#> 3                           nei                     <NA>
#> 4                            ja                      nei
#> 5                            ja                      nei
#> 6                           nei                     <NA>

 x <- data.frame(
      Indikasjon = c("Annet", NA, "NSTEMI", "NSTEMI", "NSTEMI", "NSTEMI"),
      Regtype = c("Primær", "Primær", "Sekundær", rep("Primær", 3)),
      Innkomstarsak = rep(c("Brystesmerter", "Øvrig", "Dyspne"), 2),
      Hastegrad = rep("Subakutt", 6),
      OverflyttetFra = c("Annen  avdeling på sykehuset",
                         rep("Annet sykehus", 2),
                         rep("Omdirigert ambulanse", 3)),
      ventetid_nstemi_timer = c(0, 0, 5, 100, NA, -0.1))
noric::ki_nstemi_utredet_innen24t(df_ap = x)

#>   Indikasjon  Regtype Innkomstarsak Hastegrad               OverflyttetFra
#> 1      Annet   Primær Brystesmerter  Subakutt Annen  avdeling på sykehuset
#> 2       <NA>   Primær         Øvrig  Subakutt                Annet sykehus
#> 3     NSTEMI Sekundær        Dyspne  Subakutt                Annet sykehus
#> 4     NSTEMI   Primær Brystesmerter  Subakutt         Omdirigert ambulanse
#> 5     NSTEMI   Primær         Øvrig  Subakutt         Omdirigert ambulanse
#> 6     NSTEMI   Primær        Dyspne  Subakutt         Omdirigert ambulanse
#>   ventetid_nstemi_timer indik_nstemi_angio_innen24t_data
#> 1                   0.0                              nei
#> 2                   0.0                              nei
#> 3                   5.0                              nei
#> 4                 100.0                               ja
#> 5                    NA                              nei
#> 6                  -0.1                               ja
#>   indik_nstemi_angio_innen24t
#> 1                        <NA>
#> 2                        <NA>
#> 3                        <NA>
#> 4                         nei
#> 5                        <NA>
#> 6           ugyldig/manglende

noric::ki_nstemi_utredet_innen72t(df_ap = x)

#>   Indikasjon  Regtype Innkomstarsak Hastegrad               OverflyttetFra
#> 1      Annet   Primær Brystesmerter  Subakutt Annen  avdeling på sykehuset
#> 2       <NA>   Primær         Øvrig  Subakutt                Annet sykehus
#> 3     NSTEMI Sekundær        Dyspne  Subakutt                Annet sykehus
#> 4     NSTEMI   Primær Brystesmerter  Subakutt         Omdirigert ambulanse
#> 5     NSTEMI   Primær         Øvrig  Subakutt         Omdirigert ambulanse
#> 6     NSTEMI   Primær        Dyspne  Subakutt         Omdirigert ambulanse
#>   ventetid_nstemi_timer indik_nstemi_angio_innen72t_data
#> 1                   0.0                              nei
#> 2                   0.0                              nei
#> 3                   5.0                              nei
#> 4                 100.0                               ja
#> 5                    NA                              nei
#> 6                  -0.1                               ja
#>   indik_nstemi_angio_innen72t
#> 1                        <NA>
#> 2                        <NA>
#> 3                        <NA>
#> 4                         nei
#> 5                        <NA>
#> 6           ugyldig/manglende

 x <- data.frame(
      AvdRESH = 1:6,
      Indikasjon = c("Annet", rep("STEMI",5)),
      Regtype = c("Primær", "Primær", "Sekundær", rep("Primær", 3)),
      GittTrombolyse = rep("Nei", 6),
      Hastegrad = rep("Akutt", 6),
      HLRForSykehus = rep("Nei", 6),
      ProsedyreType = rep("Angio + PCI", 6),
      BeslutningsutlosendeEKG = rep("Prehospitalt", 6),
      ventetid_stemi_min = c(-10, 20, 110, 120, 1150, 1480))
noric::ki_stemi_pci_innen120min(df_ap = x)

#>   AvdRESH Indikasjon  Regtype GittTrombolyse Hastegrad HLRForSykehus
#> 1       1      Annet   Primær            Nei     Akutt           Nei
#> 2       2      STEMI   Primær            Nei     Akutt           Nei
#> 3       3      STEMI Sekundær            Nei     Akutt           Nei
#> 4       4      STEMI   Primær            Nei     Akutt           Nei
#> 5       5      STEMI   Primær            Nei     Akutt           Nei
#> 6       6      STEMI   Primær            Nei     Akutt           Nei
#>   ProsedyreType BeslutningsutlosendeEKG ventetid_stemi_min
#> 1   Angio + PCI            Prehospitalt                -10
#> 2   Angio + PCI            Prehospitalt                 20
#> 3   Angio + PCI            Prehospitalt                110
#> 4   Angio + PCI            Prehospitalt                120
#> 5   Angio + PCI            Prehospitalt               1150
#> 6   Angio + PCI            Prehospitalt               1480
#>   indik_stemi_pci_innen2t_data indik_stemi_pci_innen2t
#> 1                          nei                    <NA>
#> 2                           ja                      ja
#> 3                          nei                    <NA>
#> 4                           ja                      ja
#> 5                           ja                     nei
#> 6                           ja       ugyldig/manglende